requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
fusepy==3.0.1
html5lib==1.1
htmllistparse==0.6.0
idna==3.4
inflect==6.0.0
minio==7.1.12
networkx==2.8.7
numpy==1.23.3
Pillow==9.2.0
pydantic==1.10.2
PyYAML==6.0
requests==2.28.1
scipy==1.9.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
webencodings==0.5.1
